requirements.txt
# INDERA Vision Dashboard — Python dependencies
flask>=3.0
torch>=2.2
transformers>=4.40
accelerate>=0.29          # for device_map="auto"
pillow>=10.0
opencv-python>=4.9
python-dotenv>=1.0
requests>=2.31
